fix: modify subscription card component
